Enrich Raw fields when possible in decode_xml
